+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @tryabby/core
 
+## 5.0.1
+
+### Patch Changes
+
+- add fallback values to remote config
+
 ## 5.0.0
 
 ### Major Changes